+++ test/circuits.test.mjs
@@ -4,8 +4,8 @@ import chaiHttp from 'chai-http';
 import chaiAsPromised from 'chai-as-promised';
 import config from 'config';
 import logger from 'common-files/utils/logger.mjs';
-import Nf3 from '../../cli/lib/nf3.mjs';
-import { expectTransaction, Web3Client } from '../utils.mjs';
+import Nf3 from '../cli/lib/nf3.mjs';
+import { expectTransaction, Web3Client } from './utils.mjs';
 
 // so we can use require with mjs file
 const { expect } = chai;
